use the same uri construction method for filecontents and editfile tools
